Modify folder structure :art: Tabs -> Spaces
