+++ src/Main.cpp
@@ -1,6 +1,6 @@
 #include "Game.hpp"
 
 int main() {
-	Game game;
-	game.run();
+    Game game;
+    game.run();
 }
